Duplicate ADJUST to a new layer (SPACE2)
Create a new layer which is currently identical to ADJUST, called
SPACE2, due to it being reachable from my second space-key. It's for
testing out if it's fine to have so many layers close to each other. If
it works, I will probably remove the ADJUST layer, as pressing down two
keys at the same time is a tiny bit annoying unless strictly necessary.